Fix: Complete Integer to Long type conversion - All build errors resolved
- Fixed RoomDTO capacity field type (Integer -> Long)
- Added capacity type casting in AdminController getAllRooms()
- All compilation errors resolved (0 errors)
- Build successful

Changes:
- RoomDTO.java: capacity field Integer -> Long
- AdminController.java: Added longValue() cast for room capacity mapping

+++ src/main/java/com/team5/demo/dto/ConferenceDTO.java
@@ -16,7 +16,7 @@ public class ConferenceDTO {
     }
 
     // Getters and Setters
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
